requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==1.10.3
attrs==24.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
compressed_rtf==1.0.6
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.21.1
numpy==2.2.1
olefile==0.47
openai==1.59.6
packaging==24.2
pandas==2.2.3
pdfminer.six==20191110
pillow==11.1.0
protobuf==5.29.3
pyarrow==18.1.0
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-pptx==0.6.23
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.12.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.8.1
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.3.0
watchdog==6.0.0
xlrd==1.2.0
XlsxWriter==3.2.0
